requirements.txt
Flask==3.1.0
mercadopago==2.2.3
python-decouple==3.8
